requirements.txt
# --- Core Framework ---
langchain>=0.1.16
langchain-core>=0.1.42
langchain-community>=0.0.32
langchain-experimental>=0.0.56
langchain-openai>=0.1.3

# --- Graph Database (Neo4j) ---
langchain-neo4j>=0.1.0
neo4j>=5.19.0

# --- Vector Database (Naive RAG) ---
langchain-chroma>=0.1.2
chromadb>=0.4.24

# --- Document Processing (PDF) ---
pypdf>=4.2.0
tiktoken>=0.6.0

# --- Data Generation & Utils ---
Faker>=24.11.0
reportlab>=4.2.0
toml>=0.10.2
python-dotenv>=1.0.1

# --- UI & Visualization (Streamlit) ---
streamlit>=1.32.0
plotly>=5.19.0

# --- API & Deployment ---
fastapi>=0.109.0
uvicorn>=0.27.0
pydantic>=2.5.0

# --- Testing & Load Testing ---
pytest>=7.4.0
pytest-asyncio>=0.23.0
httpx>=0.25.0
locust>=2.17.0

# --- Monitoring & Logging ---
requests>=2.31.0